[renderer] rename move private renderer headers
move private headers of renderer to private/le_renderer to conform with
the general rule that private headers must be in
`private/{module_name}/` so that we don't end up with private headers
shadowing each other.
