+++ modules/le_renderer/CMakeLists.txt
@@ -9,10 +9,10 @@ depends_on_island_module(le_log)
 
 set (SOURCES "le_renderer.cpp")
 set (SOURCES ${SOURCES} "le_renderer.h")
-set (SOURCES ${SOURCES} "private/le_renderer_types.h")
-set (SOURCES ${SOURCES} "private/le_vk_enums.inl")
-set (SOURCES ${SOURCES} "private/le_resource_handle_t.inl")
-set (SOURCES ${SOURCES} "private/le_rendergraph.h")
+set (SOURCES ${SOURCES} "private/le_renderer/le_renderer_types.h")
+set (SOURCES ${SOURCES} "private/le_renderer/le_vk_enums.inl")
+set (SOURCES ${SOURCES} "private/le_renderer/le_resource_handle_t.inl")
+set (SOURCES ${SOURCES} "private/le_renderer/le_rendergraph.h")
 set (SOURCES ${SOURCES} "le_rendergraph.cpp")
 set (SOURCES ${SOURCES} "le_command_buffer_encoder.cpp")
