+++ modules/le_renderer/le_renderer.h
@@ -2,7 +2,7 @@
 #define GUARD_LE_RENDERER_H
 
 #include "le_core.h"
-#include "private/le_renderer_types.h"
+#include "private/le_renderer/le_renderer_types.h"
 
 extern const uint64_t LE_RENDERPASS_MARKER_EXTERNAL; // set in le_renderer.cpp
