+++ modules/le_renderer/private/le_renderer/generate_enums.sh
