+++ modules/le_renderer/private/le_renderer/le_resource_handle_t.inl
