+++ modules/le_swapchain_vk/CMakeLists.txt
@@ -5,6 +5,7 @@ depends_on_island_module(le_core)
 depends_on_island_module(le_window)
 depends_on_island_module(le_log)
 depends_on_island_module(le_backend_vk)
+depends_on_island_module(le_renderer)
 
 add_compile_definitions(VK_NO_PROTOTYPES)
