+++ modules/le_swapchain_vk/le_swapchain_vk.cpp
@@ -1,6 +1,6 @@
 #include "le_swapchain_vk.h"
 #include "private/le_swapchain_vk/le_swapchain_vk_common.inl"
-#include "private/le_renderer_types.h" // for swapchain_settings
+#include "private/le_renderer/le_renderer_types.h" // for swapchain_settings
 #include "le_backend_vk.h"
 #include "assert.h"
 // ----------------------------------------------------------------------
